Int to Long field change

+++ antflow-base/src/main/java/org/openoa/base/vo/BpmnConfVo.java
@@ -65,7 +65,7 @@ public class BpmnConfVo {
     /**
      * process's business party
      */
-    private Integer businessPartyId;
+    private Long businessPartyId;
 
     private Integer extraFlags;
     /**
